decouple context functions
