+++ src/context/child.spec.ts
@@ -1,0 +1,9 @@
+#!/usr/bin/env -S node --no-warnings --loader ts-node/esm
+
+import { test }    from 'tstest'
+
+import * as child   from './child.js'
+
+test('tbw', async t => {
+  t.ok(child, 'tbw')
+})
