+++ src/context/context.ts
@@ -1,0 +1,42 @@
+/**
+ *   Wechaty Open Source Software - https://github.com/wechaty
+ *
+ *   @copyright 2022 Huan LI (李卓桓) <https://github.com/huan>, and
+ *                   Wechaty Contributors <https://github.com/wechaty>.
+ *
+ *   Licensed under the Apache License, Version 2.0 (the "License");
+ *   you may not use this file except in compliance with the License.
+ *   You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+import type { AnyEventObjectExt } from './origin.js'
+
+export interface Context {
+  /**
+   * current message: only the received events which should sent to child, is a `message`
+   *
+   * current message: actor module must only process one message one time
+   *  a message will only start to be processed (send to the child)
+   *  when the child is ready for processing the next message (in its idle state)
+   */
+  message?: AnyEventObjectExt
+  /**
+   * message queue: `queue` is for storing messages.
+   *
+   * a message is an event: (external events, which should be proxyed to the child)
+   *  1. neither sent from mailbox
+   *  2. nor from child
+   *
+   * TODO: Huan(202201): use yocto-queue to replace the array for better performance under high load
+   */
+  queue: AnyEventObjectExt[]
+  index: number // current message index in queue
+}
