+++ src/context/mod.ts
@@ -1,1 +1,8 @@
-export * as contexts from './contexts.js'
+export type { Context }     from './context.js'
+export { initialContext }   from './initial-context.js'
+export * as assign  from './assign.js'
+export * as child   from './child.js'
+export * as conds   from './conds.js'
+export * as origin  from './origin.js'
+export * as queue   from './queue.js'
+export * as send    from './send.js'
